pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["make87*"]

[project]
name = "make87"
version = "0"
license = { text = "make87 Platform SDK License" }
authors = [
    { name = "Nisse Knudsen", email = "nisse@make87.com" },
    { name = "Phillip Thomas", email = "phillip@make87.com" }
]
dependencies = [
    "setuptools>=80.7.1",
    "wheel>=0.45.1",
    "pydantic>=2.11.4,<3.0.0",
    "black>=25.1.0",
    "pytest>=8.3.5",
]
readme = "README.md"
requires-python = ">=3.9,<3.13"

[project.urls]
Homepage = "https://www.make87.com"
Source = "https://github.com/make87/make87-python"

[project.optional-dependencies]
zenoh = [
    "eclipse-zenoh>=1.4,<1.5",
]
protobuf = [
    "protobuf>=4.22,<7.0",
]
yaml = [
    "PyYAML>=6.0,<7.0",
]
storage = [
    "s3path",
]
dev = [
    "pre-commit",
    "pytest",
    "pytest-mock",
    "ruff>=0.11.10",
    "make87_messages",
    "datamodel-code-generator==0.30.1",
]
